requirements.txt
alembic==1.10.3
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
